requirements.txt
torch
torchvision
torchaudio
fastapi
uvicorn
pillow
tqdm
scikit-learn
